requirements.txt
Django==4.2.7
Pillow==10.1.0
python-decouple==3.8
django-extensions==3.2.3
django-crispy-forms==2.1
crispy-bootstrap5==0.7
django-phonenumber-field==7.2.0
phonenumbers==8.13.25
django-ckeditor==6.7.0
django-colorfield==0.10.1
django-mptt==0.15.0
django-taggit==4.0.0